// src/server/schemas/auth.schema.ts
import { z } from "zod";

export const loginSchema = z.object({
    email: z.email(),
    password: z.string().min(8),
});

export const registerSchema = z.object({
    email: z.email(),
    password: z.string().min(8),
    name: z.string().min(1),
});

export type LoginInput = z.infer<typeof loginSchema>;
export type RegisterInput = z.infer<typeof registerSchema>;